+++ lib/model/response_model.dart
@@ -4,13 +4,11 @@ import 'dart:convert' hide json;
 
 import 'package:freezed_annotation/freezed_annotation.dart';
 
-import 'send_model.dart';
 
 //
 part 'response_model.freezed.dart';
 part 'response_model.g.dart';
 const initConnectType = 'connected';
-typedef CustomCoverterResponseData = SendResponseModel Function(SendResponseModel nativeModel);
 
 @freezed
 class ResponseModel with _$ResponseModel {